feat: a lot of refactoring and cleanup
