+++ src/editor.scss
@@ -2,7 +2,7 @@ html {
 	--symlinks-admin-color--warning: #f0b849;
 }
 
-.c-disclosure {
+.sym-c-disclosure {
 	--c-disclosure--space: 0.5rem;
 
 	// background-color: #f0f0f0;
@@ -63,7 +63,7 @@ html {
 	}
 }
 
-.c-symlink-editor {
+.sym-c-symlink-editor {
 	$self: ".c-symlink-editor";
 
 	&__type .components-input-control__container {
@@ -79,18 +79,11 @@ html {
 		}
 	}
 
-	&__warning-icon {
-		font-size: inherit;
-		width: 1em;
-		height: 1em;
-		line-height: inherit;
-	}
-
 	&__delete {
 	}
 }
 
-.c-symlinks-sidebar {
+.sym-c-symlinks-sidebar {
 	&__list {
 		> *:not(:first-child) {
 			margin-top: 0.5rem;
@@ -98,9 +91,16 @@ html {
 	}
 }
 
-.c-warning-label {
+.sym-c-warning-label {
 	color: #1e1e1e;
 	background-color: var(--symlinks-admin-color--warning);
 	border-radius: 3px;
 	padding: 1px 0.375rem;
+}
+
+.sym-c-inline-icon {
+	font-size: inherit;
+	width: 1em;
+	height: 1em;
+	line-height: inherit;
 }
