+++ src/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { Button } from "@wordpress/components";
+import { Button, Dashicon } from "@wordpress/components";
 import { useState } from "@wordpress/element";
 import { PluginDocumentSettingPanel } from "@wordpress/edit-post";
 import { registerPlugin } from "@wordpress/plugins";
@@ -15,7 +15,7 @@ import "./store-find-post-by";
 import "./editor.scss";
 
 const PluginDocumentSettingPanelSymlinks = () => {
-	const baseClass = "c-symlinks-sidebar";
+	const baseClass = "sym-c-symlinks-sidebar";
 
 	const currentPost = useSelect((select) =>
 		select("core/editor").getCurrentPost()
@@ -24,31 +24,6 @@ const PluginDocumentSettingPanelSymlinks = () => {
 	const symlinks = useSelect(function (select) {
 		return select("core/editor").getEditedPostAttribute("meta")["symlinks"];
 	}, []);
-
-	console.log(symlinks);
-
-	// load from post meta, external table, something
-	/*const [symlinks, setSymlinks] = useState([
-		{
-			// first symlink is a simple alternate URL at the same place
-			type: "slug",
-			slug: "courses/mgmt-301-af",
-		},
-		{
-			// second symlink is set to "parent", adding a prefix of the slug of post id of 412 to the real slug
-			type: "parent",
-			parent: 2,
-		},
-		{
-			// third symlink is set to "parent-slug" using a custom slug AND adding a parent prefix
-			type: "parent-slug",
-			slug: "mgmt-301-yo",
-			parent: 1286,
-		},
-		{
-			type: "parent-slug",
-		},
-	]);*/
 
 	const panelTitle = `Symlinks${
 		symlinks.length > 0 ? ` (${symlinks.length})` : ""
@@ -80,14 +55,11 @@ const PluginDocumentSettingPanelSymlinks = () => {
 							editPost({
 								meta: { symlinks: newSymlinks },
 							});
-
-							// setSymlinks(newSymlinks);
 						}}
 					/>
 				))}
 				<Button
 					variant="secondary"
-					icon="plus"
 					className={`${baseClass}__add`}
 					onClick={() => {
 						const newSymlinks = [
@@ -101,8 +73,6 @@ const PluginDocumentSettingPanelSymlinks = () => {
 						editPost({
 							meta: { symlinks: newSymlinks },
 						});
-
-						// setSymlinks(newSymlinks);
 					}}
 				>
 					Add Symlink
